+++ src/game_types.rs
@@ -1,0 +1,71 @@
+pub struct GameHistory {
+    pub board_states: Vec<BoardState>,
+}
+impl GameHistory {
+    pub fn new_game() -> tetra::Result<GameHistory> {
+        // let board_states = vec![BoardState::default_board()?];
+        let board_states = vec![BoardState::test_board_1()?];
+        Ok(GameHistory { board_states })
+    }
+}
+pub struct BoardState {
+    pub pieces: Vec<Piece>,
+}
+impl BoardState {
+    fn test_board_1() -> tetra::Result<BoardState> {
+        let mut pieces = Vec::new();
+        let mut p = |i| pieces.push(i);
+        p(Piece(1, 4, PieceType::KNIGHT, PlayerColor::WHITE));
+        p(Piece(6, 4, PieceType::KING, PlayerColor::WHITE));
+        p(Piece(5, 3, PieceType::ROOK, PlayerColor::WHITE));
+        p(Piece(3, 1, PieceType::QUEEN, PlayerColor::WHITE));
+        p(Piece(6, 2, PieceType::KING, PlayerColor::BLACK));
+        p(Piece(4, 4, PieceType::BISHOP, PlayerColor::BLACK));
+        p(Piece(6, 6, PieceType::PAWN, PlayerColor::BLACK));
+        Ok(BoardState { pieces })
+    }
+    #[allow(dead_code)]
+    fn default_board() -> tetra::Result<BoardState> {
+        let mut pieces = Vec::new();
+        let mut p = |i| pieces.push(i);
+        for i in 0..8 {
+            p(Piece(i, 1, PieceType::PAWN, PlayerColor::WHITE));
+        }
+        p(Piece(0, 0, PieceType::ROOK, PlayerColor::WHITE));
+        p(Piece(7, 0, PieceType::ROOK, PlayerColor::WHITE));
+        p(Piece(2, 0, PieceType::BISHOP, PlayerColor::WHITE));
+        p(Piece(5, 0, PieceType::BISHOP, PlayerColor::WHITE));
+        p(Piece(1, 0, PieceType::KNIGHT, PlayerColor::WHITE));
+        p(Piece(6, 0, PieceType::KNIGHT, PlayerColor::WHITE));
+        p(Piece(4, 0, PieceType::KING, PlayerColor::WHITE));
+        p(Piece(3, 0, PieceType::QUEEN, PlayerColor::WHITE));
+        for i in 0..8 {
+            p(Piece(i, 6, PieceType::PAWN, PlayerColor::BLACK));
+        }
+        p(Piece(0, 7, PieceType::ROOK, PlayerColor::BLACK));
+        p(Piece(7, 7, PieceType::ROOK, PlayerColor::BLACK));
+        p(Piece(2, 7, PieceType::BISHOP, PlayerColor::BLACK));
+        p(Piece(5, 7, PieceType::BISHOP, PlayerColor::BLACK));
+        p(Piece(1, 7, PieceType::KNIGHT, PlayerColor::BLACK));
+        p(Piece(6, 7, PieceType::KNIGHT, PlayerColor::BLACK));
+        p(Piece(4, 7, PieceType::KING, PlayerColor::BLACK));
+        p(Piece(3, 7, PieceType::QUEEN, PlayerColor::BLACK));
+        Ok(BoardState { pieces })
+    }
+}
+#[derive(Clone, Copy)]
+pub struct Piece(pub i8, pub i8, pub PieceType, pub PlayerColor);
+#[derive(PartialEq, Clone, Copy, Debug)]
+pub enum PieceType {
+    PAWN,
+    ROOK,
+    BISHOP,
+    KNIGHT,
+    KING,
+    QUEEN,
+}
+#[derive(PartialEq, Clone, Copy, Debug)]
+pub enum PlayerColor {
+    BLACK,
+    WHITE,
+}
